models folder refactor
