+++ packages/reporting-server/models/tortoise_models/container.py
@@ -1,0 +1,6 @@
+from tortoise import fields, models
+
+
+class Container(models.Model):
+    id = fields.IntField(pk=True)
+    name = fields.TextField()
