+++ packages/reporting-server/models/tortoise_models/health.py
@@ -1,7 +1,6 @@
 from enum import Enum
 
 from tortoise import fields, models
-from tortoise.contrib.pydantic import pydantic_model_creator
 
 
 class HealthStatusEmun(Enum):
@@ -16,6 +15,3 @@ class HealthStatus(models.Model):
     health_status = fields.CharField(max_length=25, null=True)
     health_message = fields.TextField(null=True)
     created = fields.DatetimeField(auto_now_add=True)
-
-
-HealthStatus_Pydantic = pydantic_model_creator(HealthStatus, name="HealthStatus")
