+++ packages/reporting-server/parsers/health_parser.py
@@ -1,6 +1,6 @@
 import json
 
-from models.health import HealthStatus
+from models.tortoise_models.health import HealthStatus
 
 
 async def health_status_parser(fullstring: str, health_device: str):
