+++ packages/reporting-server/rest_server/repositories/log_creation_handler.py
@@ -1,9 +1,9 @@
 import logging
 
 from dependencies import logger
-from models.auth_events import AuthEvents
-from models.container import Container
-from models.raw_log import RawLog
+from models.tortoise_models.auth_events import AuthEvents
+from models.tortoise_models.container import Container
+from models.tortoise_models.raw_log import RawLog
 from parsers.auth_event_parser import auth_event_parser
 from parsers.log_type_parser import get_log_type
