+++ packages/reporting-server/rest_server/repositories/parser_dispatcher.py
@@ -1,10 +1,10 @@
-from models.dispenser_state import DispenserState
-from models.door_state import DoorState
-from models.fleet_state import FleetState
-from models.health import HealthStatus
-from models.ingestor_state import IngestorState
-from models.lift_state import LiftState
-from models.task_summary import TaskSummary
+from models.tortoise_models.dispenser_state import DispenserState
+from models.tortoise_models.door_state import DoorState
+from models.tortoise_models.fleet_state import FleetState
+from models.tortoise_models.health import HealthStatus
+from models.tortoise_models.ingestor_state import IngestorState
+from models.tortoise_models.lift_state import LiftState
+from models.tortoise_models.task_summary import TaskSummary
 from parsers.dispenser_state_parser import dispenser_state_parser
 from parsers.doors_state_parser import doors_state_parser
 from parsers.fleet_state_parser import fleet_state_parser
