+++ packages/reporting-server/rest_server/repositories/report/fleet_state.py
@@ -2,7 +2,8 @@ from datetime import timezone
 from typing import Optional
 
 from dateutil import parser
-from models.fleet_state import FleetState, FleetState_Pydantic
+from models.pydantic_models import LiftState_Pydantic
+from models.tortoise_models.fleet_state import FleetState
 
 
 async def get_fleet_state(
@@ -23,6 +24,6 @@ async def get_fleet_state(
         to_log_utc_time = to_log_local_time.astimezone(timezone.utc)
         query["created__lt"] = to_log_utc_time
 
-    return await FleetState_Pydantic.from_queryset(
+    return await LiftState_Pydantic.from_queryset(
         FleetState.filter(**query).offset(offset).limit(limit).order_by("-created")
     )
