+++ packages/reporting-server/rest_server/repositories/test_log_creation_handler.py
@@ -3,9 +3,9 @@
 import unittest
 
 from fastapi.testclient import TestClient
-from models import RawLog
-from models.auth_events import AuthEvents
-from models.container import Container
+from models.tortoise_models import RawLog
+from models.tortoise_models.auth_events import AuthEvents
+from models.tortoise_models.container import Container
 from rest_server.__mocks__ import raw_data
 from rest_server.app import get_app
 from tortoise import Tortoise
